Moved RestController class from FullRegistration module to
CustomerRegistration parent
