+++ CustomerRegistration/pom.xml
@@ -142,6 +142,8 @@
 			<version>1.0-SNAPSHOT</version>
 		</dependency>
 	</dependencies>
+	
+	
 	<repositories>
 		<repository>
 			<id>spring-milestone</id>
