--- CustomerRegistration/src/main/java/com/gogenie/customer/registration/App.java
@@ -1,13 +1,0 @@
-package com.gogenie.customer.registration;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
