+++ fullregistration/src/main/java/com/gogenie/customer/fullregistration/service/impl/FullRegistrationServiceImpl.java
@@ -3,15 +3,19 @@ package com.gogenie.customer.fullregistration.service.impl;
 import javax.inject.Inject;
 import javax.inject.Named;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import com.gogenie.customer.fullregistration.dao.FullRegistrationDAO;
 import com.gogenie.customer.fullregistration.model.RegistrationRequest;
 import com.gogenie.customer.fullregistration.model.RegistrationResponse;
 import com.gogenie.customer.fullregistration.service.FullRegistrationService;
 
 @Named
+@Service
 public class FullRegistrationServiceImpl implements FullRegistrationService {
 
-	@Inject
+	@Autowired
 	FullRegistrationDAO fullRegistrationDao;
 
 	@Override
